+++ src/modules/adsb-feeder/filesystem/root/opt/adsb/adsb-setup/utils/data.py
@@ -20,6 +20,8 @@ class Data:
     env_file_path = config_path / ".env"
     version_file = data_path / "adsb.im.version"
     secure_image_path = data_path / "adsb.im.secure_image"
+    # hardcoded path in net-or-hotspot, also hardcode here
+    hotspot_disabled_path = Path("/opt/adsb/adsb.im.hotspot_disabled")
     is_feeder_image = True
     _env_by_tags_dict: dict[tuple[str, ...], Env] = field(default_factory=dict[tuple[str, ...], Env])